Renamed files, added `web_remaps` table
